package & docs update
Just a couple of changes I made while setting up locally:

`npm install` fails (Ubuntu 16.4 LTS, node 4.3.1) on the package `node-pre-gyp`, which is a dependency of the pacakge `grunt-node-inspector`. Updating to a later version of the package works.

I also fixed a couple of typos in the docs, and added a `setup` section.